+++ macrobond_data_api/web/web_types/values_response.py
@@ -8,7 +8,7 @@ from .entity_response import EntityResponse
 class ValuesResponse(EntityResponse):
     """A series"""
 
-    values: Optional[List[float]]
+    values: Optional[List[int]]
     """
     The values of the series or not specified if there was an error.
     Missing values are represented by null.
